requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
python-dotenv
requests
starlette
itsdangerous
pydantic-settings
